requirements.txt
jupyterlab==3.2.5
torch
torchtyping
einops
torchvision
torchtext
numpy
transformers
sentencepiece
unidecode
jupyter_kernel_gateway
ipykernel
gin-config
pytest
matplotlib
plotly
sklearn
black
comet_ml
mediapy
numpy-stl
traitlets==5.1.1
git+https://github.com/google/gin-config.git@e518c4ec7755a3e5da973e894ab23cc80c6665ed#egg=gin-config
